[gui] extract common logic from disk and ram sliders
